Revert "Reset new fork to old fork"
This reverts commit eb2758b5344903abacc2e6d0d82c64bf0f5fa873.

Something in the new fork is causing a topline metrics regression. We're
not sure what it is, so we're going to split it into steps and bisect.

As a first step, this resets the new fork back to the contents of the
old fork. We will land this to confirm that the fork infra itself is
not causing a regression.